requirements.txt
black
certifi
charset-normalizer
click
colorama
filelock
flake8
fsspec
idna
Jinja2
MarkupSafe
matplotlib
mccabe
mpmath
mypy-extensions
networkx
nibabel
numpy<2
opencv-python
opencv-contrib-python
packaging
pandas
pathspec
pillow
platformdirs
pycodestyle
pyflakes
python-dateutil
pytz
ruamel.yaml
requests
scikit-image
scikit-learn
setuptools
six
sympy
tensorboardX
torch
torchinfo
torchmetrics
torchmetrics[detection]
torchvision
tqdm
typing_extensions
tzdata
urllib3
wheel